youth_04_08/pyproject.toml
[project]
name = "youth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
